+++ NEWS.md
@@ -95,6 +95,8 @@ The changes are reflected in the installation section of the handbook.
 - Installation script: increase download timeout to avoid (unclear) errors when the script is downloading large file (issue #76).
 - Fixed: `checkFeatures()`/`checkComponents()`: disabling a feature/featureGroup in a sorted table would lead to wrong selections
 - `annotateSuspects()`: log if the suspect formula/compound data could not be matched with feature annotations
+- Fixed: `chromPeaks()` from `xcms` was somethimes not found (issue #68) 
+- `plotGraph()` method for transformation products gained `width` and `height` arguments.
 
 
 # patRoon 2.1